+++ src/samigo/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/listener/select/SelectActionListener.java
@@ -1,6 +1,6 @@
 /**********************************************************************************
  * $URL: https://source.sakaiproject.org/svn/sam/trunk/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/listener/select/SelectActionListener.java $
- * $Id: SelectActionListener.java 107566 2012-04-25 14:16:22Z david.horwitz@uct.ac.za $
+ * $Id: SelectActionListener.java 121258 2013-03-15 15:03:36Z ottenhoff@longsight.com $
  ***********************************************************************************
  *
  * Copyright (c) 2004, 2005, 2006, 2007, 2008, 2009 The Sakai Foundation
@@ -66,7 +66,7 @@ import org.sakaiproject.util.ResourceLoader;
  * <p>Purpose:  this module creates the lists of published assessments for the select index
  * <p>Description: Sakai Assessment Manager</p>
  * @author Ed Smiley
- * @version $Id: SelectActionListener.java 107566 2012-04-25 14:16:22Z david.horwitz@uct.ac.za $
+ * @version $Id: SelectActionListener.java 121258 2013-03-15 15:03:36Z ottenhoff@longsight.com $
  */
 
 public class SelectActionListener
@@ -316,33 +316,33 @@ public class SelectActionListener
     
     String lastPublishedAssessmentId = "";
     HashMap averageScoreMap = new HashMap();
-    float totalScores= 0f;
+    double totalScores= 0d;
 	int totalSubmissions= 0;
-	float averageScore = 0f;
+	double averageScore = 0d;
     
 	for (int i = 0; i < averageScoreAssessmentGradingList.size(); i++)
 	{
 		DeliveryBeanie db = (DeliveryBeanie) averageScoreAssessmentGradingList.get(i);
 		if ((lastPublishedAssessmentId != null && lastPublishedAssessmentId.equals(db.getAssessmentId())) || averageScoreAssessmentGradingList.size() == 1) {
-			totalScores += Float.parseFloat(db.getFinalScore());
+			totalScores += Double.parseDouble(db.getFinalScore());
 			totalSubmissions++;
 			if (i == averageScoreAssessmentGradingList.size() - 1) {
 				averageScore = totalScores/totalSubmissions;
-				averageScoreMap.put(db.getAssessmentId(), Float.valueOf(averageScore));
+				averageScoreMap.put(db.getAssessmentId(), Double.valueOf(averageScore));
 			}
 		}
 		else {
 			if (i > 0) {
 				averageScore = totalScores/totalSubmissions;
-				averageScoreMap.put(lastPublishedAssessmentId, Float.valueOf(averageScore));
+				averageScoreMap.put(lastPublishedAssessmentId, Double.valueOf(averageScore));
 			}
 			lastPublishedAssessmentId = db.getAssessmentId(); 
-			totalScores = Float.parseFloat(db.getFinalScore());
+			totalScores = Double.parseDouble(db.getFinalScore());
 			totalSubmissions = 1;
 			
 			if (i == averageScoreAssessmentGradingList.size() - 1) {
 				averageScore = totalScores/totalSubmissions;
-				averageScoreMap.put(db.getAssessmentId(), Float.valueOf(averageScore));
+				averageScoreMap.put(db.getAssessmentId(), Double.valueOf(averageScore));
 			}
 		}
 	}
