+++ src/samigo/samigo-qti/src/java/org/sakaiproject/tool/assessment/qti/helper/item/ItemHelper20Impl.java
@@ -1,6 +1,6 @@
 /**********************************************************************************
  * $URL: https://source.sakaiproject.org/svn/sam/trunk/samigo-qti/src/java/org/sakaiproject/tool/assessment/qti/helper/item/ItemHelper20Impl.java $
- * $Id: ItemHelper20Impl.java 106463 2012-04-02 12:20:09Z david.horwitz@uct.ac.za $
+ * $Id: ItemHelper20Impl.java 121258 2013-03-15 15:03:36Z ottenhoff@longsight.com $
  ***********************************************************************************
  *
  * Copyright (c) 2003, 2004, 2005, 2006, 2008, 2009 The Sakai Foundation
@@ -47,7 +47,7 @@ import org.sakaiproject.tool.assessment.qti.helper.AuthoringXml;
  * <p>Organization: Sakai Project</p>
  * <p>Version for QTI 2.0 item XML, significant differences between 1.2 and 2.0</p>
  * @author Ed Smiley esmiley@stanford.edu
- * @version $Id: ItemHelper20Impl.java 106463 2012-04-02 12:20:09Z david.horwitz@uct.ac.za $
+ * @version $Id: ItemHelper20Impl.java 121258 2013-03-15 15:03:36Z ottenhoff@longsight.com $
  */
 
 public class ItemHelper20Impl extends ItemHelperBase
@@ -73,12 +73,12 @@ public class ItemHelper20Impl extends ItemHelperBase
    * @param score
    * @param itemXml
    */
-  public void addMaxScore(Float score, Item itemXml)
+  public void addMaxScore(Double score, Item itemXml)
   {
     // normalize if null
     if (score == null)
     {
-      score =  Float.valueOf(0);
+      score =  Double.valueOf(0);
     }
     // set the responseElse baseValue, if it exists
     String xPath =
@@ -98,12 +98,12 @@ public class ItemHelper20Impl extends ItemHelperBase
    * @param score
    * @param itemXml
    */
-  public void addMinScore(Float score, Item itemXml)
+  public void addMinScore(Double score, Item itemXml)
   {
     // normalize if null
     if (score == null)
     {
-      score =  Float.valueOf(0);
+      score =  Double.valueOf(0);
     }
     // first, set the outcomeDeclaration defaultValue, if it exists
     String xPath =
